1972 Ping Pong Clean room FPGA
This project is a faithful, transistor-level recreation of Atari's Pong (1972), the first commercially successful video game, implemented on a DE1-SoC FPGA board. This is a clean-room design, developed solely based on observations of the original game's behavior and publicly available information, without access to Atari's original schematics, all the functionality and score tracking of the original product present.